BUG#36657936 rescan() on Replica Cluster: endless loop regarding group view uuid
When a ClusterSet is composed of clusters running different versions,
particularly when the Primary Cluster is on version >= 8.3.0 and the
Replica Clusters are on versions < 8.3.0, it is expected that the
Primary Cluster will keep the default value for
'group_replication_view_change_uuid'. However, Replica Clusters on
versions < 8.3.0 require a value to be set, as it's mandatory for
ClusterSet functionality.

Running .rescan() on one of the Replica Clusters resulted in a warning
indicating that 'group_replication_view_change_uuid' is not configured
but is required.

The root cause of this issue was that the check for this setting was
incorrectly performed on the Primary Cluster instead of the target
Replica Cluster.

This patch corrects that behavior.

Change-Id: Iaa503e08e0ed3e46bb2b07a1d0c9dc92861ff162