BUG#36054619 BUG#36682741 [postfix]
When creating a Replica Cluster with .create_replica_cluster() and
specifying the 'cloneDonor' option, recovery validations were
incorrectly being performed against the primary member of the Primary
Cluster instead of the instance passed to 'cloneDonor'.

This patch corrects the behavior by ensuring that, when the
'cloneDonor' option is used, validations are performed against the
specified instance. If the option is not provided, the validations
default to the primary member of the Primary Cluster since that's the
default instance used as donor for clone-based recovery.

Additionally, the output text has been updated in the mixed_version
tests to reflect the other changes of the previous patch.

Change-Id: Ica5fdf6ae9fa1ef9ecb9264a58c4371ab020afaa