BUG#34582616 util.importTable barking about X protocol, dump/export ones don't
The util.importTable() was throwing an error when the global shell
session was using the X protocol.

With this fix, error is no longer reported, instead the existing X
protocol session is used to establish a classic connection.

Additional changes
------------------

Added support for PAR URLs to `util.exportTable()` and
`util.importTable()` utilities.

Options in the output of a CLI help call, i.e.:
  mysqlsh -- util dump-instance --help
are listed alphabetically.

Internal refactoring, storage options have been extracted to a separate
class.

Change-Id: I6da81d57b73ee6bf71fa1b7c9b22c7cc0521d925