BUG#37158908 Dump may hang when executing a binlog consistency check
When running a dump on an active instance with the `consistent` option
set to true, using an account which lacks privileges to execute FLUSH
TABLES WITH READ LOCK, shell runs some additional consistency checks and
validates executed statements using the information from binlogs.

If this validation was triggered because binlog file/position has
changed and more than one binlog file needed to be examined for executed
statements, Shell could hang while running this consistency check.

Due to a bug, all binlog events were checked, and some of them had an
empty Info field, which caused Shell to enter an infinite loop.

As a fix, only non-empty Query events are checked. Additionally, GRANT
and REVOKE statements are also flagged as unsafe, if consistency check
detects them, an exception is thrown.

Change-Id: Ife4c7df74b02c596c65acb2b9fe76e46a54fea33