BUG#37216767 The `maxRate` dump option may not limit the throughput
The `Rate_limit m_rate_limit` variable used by the dumper to throttle
the dump was initialized when the dump started. If there was a time gap
between start of the dump and start of the data dump (i.e. when there
was a lot of data that needed to be chunked), the throttling would be
off and dump was not throttled for some period of time.

As a fix, the scope of `Rate_limit` variable is now limited to the
`dump_table_data()` method, meaning the information required to throttle
the transfer is collected only when data is dumped.

Change-Id: I3391f4b9eec7b00b0a2d23c2766a4d0e7a299b7b