BUG#37334759 BUG#37335360 Replication topology detection
BUG#37334759 Replication topology detection on former members: error out or not?

Topology detection of dumpBinlogs() has been updated, now it works as
follows:
 - instance is the same as the one in the previous dump - OK
 - instance is different than the one in the previous dump:
   - instance is not part of an InnoDB Cluster - error out
   - instance is part of an InnoDB Cluster, but it's not ONLINE - error
     out
   - instance is part of an InnoDB Cluster, it's ONLINE, but it's a
     different cluster than the one in the previous dump - error out
   - else: OK

Additional changes in util.loadBinlogs():
 - `stopBefore` no longer accepts a <file>:<position> value
 - added `stopAfter` option, which is the same as `stopBefore`, but
   stops the load after the specified transaction is executed

BUG#37335360 Read replica says source is not ONLINE, but it is

The read replica belonging to an InnoDB cluster is no longer supported,
added an explicit message to explain this.

Change-Id: I3f79f7bec1e3378215d3147a4c7fe9682764e141