BUG#37344704 ClusterSet modifies guidelines that are not active
When a ClusterSet is created, the Cluster's default guidelines should be
automatically upgraded to the default ClusterSet guideline.

However, the implementation did not account for cases where default
guidelines were modified after their creation, making them non-default.
Such guidelines should not be upgraded when the Cluster transitions to a
ClusterSet.

This patch ensures that during ClusterSet creation, only the unchanged
default guidelines are upgraded to the ClusterSet's default guidelines,
regardless of whether they are active or not.

Change-Id: I1fff57425cdc711eff78e4d4113aa49603ada781