BUG#37344823 Auto upgrade of a guideline referenced by a Router
When a ClusterSet is created, default routing guidelines are
automatically upgraded to the default ClusterSet guideline, regardless
of whether the guideline is active.

However, the default guideline created by Shell did not account for
scenarios where Routers had not been re-bootstrapped after the
ClusterSet was created. This led to broken expectations compared to the
behavior before routing guidelines, where Routers would continue
operating as if the Cluster was standalone. In such cases, Routers fail
to operate correctly due to the upgraded guideline.

This patch addresses these issues by:

  - Updating the default ClusterSet Routing Guidelines configuration to
    handle scenarios where $.server.clusterRole is not recognized. This
    is achieved by using the UNDEFINED value to account for unknown
    roles.
  - Adding a warning when upgrading guidelines and Routers are detected,
    informing users that a re-bootstrap is required for the Routers to
    recognize the updated configuration.

Change-Id: I7a47d63938707efbeff8bb083a29f7682f3638a3