BUG#37344502 setRouteOption(destinations) - allow optional quoting of class names
Quotes are currently not allowed in the "destinations" selector
values used in .add_route() and .set_route_option().

However, there is no technical reason to restrict this. For example, the
following formats should be valid:

  - ["round-robin('PrimaryClusterSecondary')"]
  - ['round-robin("Primary", "Secondary")"]
  - ["round-robin(\'Primary\', \'Secondary\')"]

This patch enhances the regex used to parse and validate destination
selectors to allow single and double quoting. Additionally, it removes
the quotes before processing the values, ensuring consistent handling.

Change-Id: I2b3f76adb2764efdcc56e4c32354b0c833ff9ede