Fix test failing on ARM platforms
Fixed test failing on ARM platforms due to out of order SQL calls.

Traces are recorded on x86-64 platform which has a different order of
parameter evaluation than ARM. When parameters are results of two or
more function calls, and each of these functions executes some kind of
SQL, traces recorded on one platform do not work on the other one.

Change-Id: Ifc7402022fdda0edf2debe1b23af56b2af8c2600