BUG#37244012 cluister.routerOptions(extended: 2), Routing Guidelines and Router 9.1
When a Router with a version < 9.2 is bootstrapped in a topology using
Routing Guidelines, the .router_options({"extended": 2}) command
incorrectly reports the guideline set as active in the topology in that
Router's configuration.

Since Routers < 9.2 do not support Routing Guidelines, this report is
wrong.

This patch ensures that Routers without Routing Guidelines support do
not include the active guideline in their configuration report.

Change-Id: I38ca054ed32abc5c2c7fcbf88e03016cb3bda0d8