BUG#37101286 Router reports: Ignoring unsupported instance <host:port>, type: 'group-member'
When upgrading the metadata schema of a ReplicaSet to version 2.2.0, the
'instances' table is not being populated correctly. This results in
MySQL Router failing to recognize the topology and rejecting connections.

Metadata model 2.2.0 introduces the 'instance_type' column in the
'instances' table with the enum values: 'group-member', 'async-member',
and 'read-replica'. During the upgrade via 'dba.upgrade_metadata()', the
command does not determine whether the topology is an InnoDB Cluster or
a ReplicaSet. Consequently, the 'instance_type' column is assigned its
default value ('group-member'), leading to incorrect metadata.

This patch ensures that the metadata upgrade process properly assigns
the correct 'instance_type' value based on the topology: 'group-member'
for Clusters and 'async-member' for ReplicaSets.

Change-Id: Id5f17257b3e0ca3c10ede47cfe40e1d6b62c52af