BUG#35282392 ReplicaSet.removeInstance() with {force: true} cannot the remove instance if the replication is in error state
When a ReplicaSet member enters the ERROR state, such as in the case of
a GTID mismatch, the .removeInstance() command fails to remove the
instance, even when the force option is specified. This behavior
contradicts the intended purpose of the force option, which is to allow
such removals.

This patch resolves the issue by ensuring that .removeInstance() can
successfully remove a ReplicaSet member in the ERROR state, provided the
force option is explicitly used.

Change-Id: I7fc85e3f66b0ca85e6f83ec913bfe97f20351d73