BUG#35914505 replicaSet.status() replicationLag 'null' is confusing
The 'replicationLag' field in replicaset.status() currently reports
'null' when replication is idle, i.e., when all transactions have been
applied. This behavior is misleading, as it may be interpreted as
replication not functioning.

This patch addresses the issue by adopting the same reasoning applied in
BUG#34002328 and ensures the following behavior:

  - If the replication connection or SQL thread is not running,
    'replicationLag' is set to 'null'.
  - When the last queued transaction matches the last applied
    transaction, or the applying transaction count is 0, indicating the
    applier queue is fully processed, 'replicationLag' is set to
    'applier_queue_applied'.

Change-Id: I1718130a61aa60a0c378ee2b0864385fd690b888