BUG#37629756  Automatically convert a file PAR to a prefix PAR
When dumping or loading using a prefix PAR, users need to make sure
that the provided URL ends with a slash, otherwise an error is
reported:

  Invalid PAR, expected: https://<namespace>.objectstorage.<region>.oci.customer-oci.com/p/<secret>/n/<namespace>/b/<bucket>/o/[<prefix>/]

The URLs of prefix PARs generated by Console do not end with slash, so
it's easy to hit this error, and it's often hard to spot the issue.

When a PAR URL which does not end with a slash (a file PAR) is provided,
i.e.:

https://<namespace>.objectstorage.<region>.oci.customer-oci.com/p/<secret>/n/<namespace>/b/<bucket>/o/<prefix>

and a prefix PAR is expected (a directory PAR):

https://<namespace>.objectstorage.<region>.oci.customer-oci.com/p/<secret>/n/<namespace>/b/<bucket>/o/<prefix>/

the missing slash is now added automatically and operation continues.

Change-Id: Ic09fc363caf062c54af46c44e32354237895460d