BUG#37408046 util.importTable() may crash when reporting an error
The util.importTable() utility importing from a single file could crash
if one of the background threads reports an error.

The flag which indicates that an exception is set was updated before
the exception had been stored. If the controlling thread attempted to
rethrow the exception after the flag was toggled, but before exception
was stored, it would cause a crash.

The race condition has been removed.

Change-Id: Icbbe2b5662babd6df3d2a3fbf65af56c10371639