Add support for Routing Guidelines version 1.1
As a follow-up to Bug#37408429 (Router), this patch introduces support
for Routing Guidelines version 1.1.

In version 1.1, Router changes how tag strings are handled. When version
1.1 is used, Router allows matching tags to preserve their type. For
version 1.0, Router assumes that tags are always passed as strings.

This patch updates the default guideline version to 1.1, adds support
for it, and ensures backward compatibility with version 1.0. To achieve
this, the auto-escaping logic for key-value types is applied only for
version 1.0, while it is skipped for version 1.1.

Change-Id: I2f38df0d167104280edc320d41504e5808a181c6