Don't show traceback in Python plugin exceptions
Python plugins that threw exceptions had the traceback included in the
exception text, which was leaking through to the UI. This was happening
in Python_context::fetch_and_clear_exception() which was returning the
full formatted exception text when only the exception message was needed.
The function was changed to return the original exception type, message
text and traceback separately, which can then be forwarded to the GUI or
displayed as needed by the UI layer.

Change-Id: I9b24d34c26464dac87a5035facf222e5dfd51bf4