BUG#37593239 util.copySchemas() may hang when target user doesn't have P_S.* SELECT privileges
When loading (or copying) a dump using an account which lacks the SELECT
privilege on performance_schema, if indexes were rebuilt
(`deferTableIndexes` set to 'all' or 'fulltext'), load operation would
hang when indexes were being rebuilt.

The function which is updating the progress did not handle query errors
and terminated early in case of any exceptions, while other threads
were waiting for progress to reach 100%.

Error handling was added.

Change-Id: I05a3703bde9402462bbaf1e6ff28775329a259d8