BUG#37399749 Switchover in InnoDB Cluster Incorrectly Requires All Members to Be Online
The AdminAPI incorrectly requires all members of an InnoDB Cluster to be
ONLINE before allowing a switchover. This requirement is too strict
because Group Replication only requires that the group has quorum.

This patch relaxes the precondition checks in
<Cluster>.set_primary_instance() to require only quorum instead of
enforcing that all members are ONLINE.

The behavior of <Cluster>.switch_to_multi_primary_mode() and
<Cluster>.switch_to_single_primary_mode() remains unchanged because
auto_increment_* values must be updated on all cluster members, making
it necessary to enforce that all members are ONLINE.

Change-Id: I58404f2d29a9f31d36aae487f6649f3b4ed363af