WL#16802 BUG#38224892 BUG#38219104 Migration of Vector Store tables
WL#16802 Dump and load support for migrating Vector Store tables from InnoDB to Lakehouse

InnoDB vector store tables
==========================

These are the InnoDB tables with at least one GenAI `VECTOR` column,
which are using the `RAPID` secondary engine.

New option added to the `util.dumpInstance()`, `util.dumpSchemas()` and `util.dumpTables()`
===========================================================================================

`lakehouseTarget`
-----------------

dictionary (default: not set) - Specifies where the data of InnoDB
vector store tables will be written. The following sub-options are
supported:

* `outputUrl` - string (required)
* `osBucketName` - string (default: not set) - Name of the OCI Object
  Storage bucket to use.
* `osNamespace` - string (default: not set) - Specifies the namespace
  where the bucket is located.
* `ociConfigFile` - string (default: not set) - Use the specified OCI
  configuration file instead of the one at the default location.
* `ociProfile` -  string (default: not set) - Use the specified OCI
  profile instead of the default one.
* `ociAuth` - string (default: not set) - Use the specified
  authentication method when connecting to the OCI.

This option can only be used when the `ocimds` option is set to true.

If this option is not given and the `ocimds` option is set to true, the
data of InnoDB vector store tables is going to be stored in a separate
directory inside of the main dump directory.

Allowed values of `outputUrl` sub-option:

* `[file://]/local/path` - local file storage
* `oci/bucket/path` - OCI Object Storage, when the `osBucketName`
  sub-option is given
* OCI Pre-Authenticated Request to a bucket or a prefix

If the `outputUrl` sub-option is set to a non-URI value (the `scheme://`
part is not present) and the `osBucketName` sub-option is not given, the
value of `outputUrl` sub-option specifies a location in the same storage
as the rest of the dump, i.e. if the `outputUrl` parameter is set to a
PAR, the `outputUrl` sub-option specifies a different prefix in that PAR.

Note: If the InnoDB vector store table data is stored in a local storage,
it's not possible convert such tables to Lakehouse tables.

New options added to the `util.loadDump()`
==========================================

`convertInnoDbVectorStore`
--------------------------

string (default: `auto`) - Specifies how to handle the InnoDB vector
store tables when loading a dump. Allowed values: `auto`, `skip`, `keep`,
`convert`.

When the `convertInnoDbVectorStore` option is set to `auto`:

* If the target instance is not a MySQL HeatWave Service instance or if
  the vector store table data is not stored in an OCI's Object Storage,
  the InnoDB vector store tables are loaded as regular tables.
* If the target instance is a MySQL HeatWave Service instance and the
  vector store table data is stored in an OCI's Object Storage, the
  InnoDB vector store tables are converted to Lakehouse tables.

When the `convertInnoDbVectorStore` option is set to `skip`:

* The InnoDB vector store tables are not loaded.

When the `convertInnoDbVectorStore` option is set to `keep`:

* The InnoDB vector store tables are loaded as regular tables.

When the `convertInnoDbVectorStore` option is set to `convert`:

* If the target instance is not a MySQL HeatWave Service instance or if
  the vector store table data is not stored in an OCI's Object Storage,
  an exception is thrown.
* If the target instance is a MySQL HeatWave Service instance and the
  vector store table data is stored in an OCI's Object Storage, the
  InnoDB vector store tables are converted to Lakehouse tables.

Note: If `convertInnoDbVectorStore` is set to `convert`, MySQL HeatWave
9.4.1 or newer is required.

`lakehouseSource`
-----------------

dictionary (default: not set) - Specifies where the data of InnoDB vector
store tables is located. The following sub-options are supported:

* `par` - string (default: not set)
* `bucket` - string (default: not set)
* `namespace` - string (default: not set)
* `region` -  string (default: not set)
* `prefix` -  string (default: not set)

Either `par` sub-option or `bucket`, `namespace`, `region` sub-options are
required.

The `prefix` sub-option is optional and may be used with `bucket`,
`namespace`, `region` sub-options.

If the `lakehouseSource` option is not set, the location of InnoDB vector
store table data is read from dump's metadata.

`heatwaveLoad`
--------------

string (default: `vector_store`) - Specifies which tables are going to be
loaded into the HeatWave engine. Allowed values: `none`, `vector_store`.

When the target instance is not a MySQL HeatWave instance, or when the
`heatwaveLoad` option is set to `none`, the `heatwaveLoad` option is
ignored.

When the `heatwaveLoad` option is set to `vector_store` and the InnoDB
vector store tables were converted to Lakehouse tables (see
`convertInnoDbVectorStore` option), they are loaded into the HeatWave
engine.

New loader errors have been added:
----------------------------------

define SHERR_LOAD_INNODB_VECTOR_STORE_UNKNOWN_LOCATION 53033
define SHERR_LOAD_INNODB_VECTOR_STORE_UNKNOWN_LOCATION_MSG              \
  "Cannot load the InnoDB based vector store tables as regular tables, " \
  "location of their data is not known."

define SHERR_LOAD_INNODB_VECTOR_STORE_NOT_MHS 53034
define SHERR_LOAD_INNODB_VECTOR_STORE_NOT_MHS_MSG                     \
  "Cannot convert the InnoDB based vector store tables to Lakehouse, " \
  "target instance is not a HeatWave instance."

define SHERR_LOAD_INNODB_VECTOR_STORE_UNSUPPORTED_MHS 53035
define SHERR_LOAD_INNODB_VECTOR_STORE_UNSUPPORTED_MHS_MSG             \
  "Cannot convert the InnoDB based vector store tables to Lakehouse, " \
  "target HeatWave instance needs to be 9.4.1 or newer."

define SHERR_LOAD_INNODB_VECTOR_STORE_NON_OCI_LOCATION 53036
define SHERR_LOAD_INNODB_VECTOR_STORE_NON_OCI_LOCATION_MSG            \
  "Cannot convert the InnoDB based vector store tables to Lakehouse, " \
  "their data is not stored in OCI's Object Storage."

BUG#38224892 util.loadDump() claims vector table import even if import fails

Vector store statistics at the end of a load were not reported correctly
if one of the tasks had failed.

BUG#38219104 When overwriting vector dump dir, integrity of old dump may be impacted

Dumper was not checking if directory specified by 'lakehouseTarget' is
empty before starting the dump, possibly overwriting its contents.

With this fix, if that directory is not empty, dump fails with an error.

Change-Id: I06005dfe2e608c8c264e53bbd3c9d26f9d2f974b