Failing openid connect windows tests fix
This commit fixes a failing openid connect test on windows that is due to javascript os.path.join using a backslash character when joing path directories - this will produce an incorrect path (escaped), when used in an SQL query.

Also this fixes the isAuthMethodSupported for auth plugins already installed on server. There was a bug with exception error code being not properly read - due to this, when queried for an already installed plugin, the isAuthMethodSupported returned false instead of true.

Change-Id: Ia7fe7f282b030dafc552f1a6f54e1f073c35f6cf