BUG#37739359 <RoutingGuideline>.show({"router": <router>}) fails if the guideline belongs to a ClusterSet
Using the 'router' option in <RoutingGuideline>.show() fails when the
guideline belongs to a ClusterSet.

The command incorrectly reports that the specified Router does not
belong to the ClusterSet, even when it does.

This issue is caused by attempting to retrieve Router metadata using
logic intended for Cluster instead of ClusterSet.

This patch corrects the behavior by applying the appropriate logic
when the guideline is associated with a ClusterSet.

Change-Id: If6e4dcf16d6e8e857c6e2a1e9b730aa67b0b97a1