BUG#37750932 Unable to import backed-up routing guideline with duplicate name
When importing a backed-up routing guideline with the
.import_routing_guideline() command, the operation fails if a guideline
with the same name already exists in the topology.

While the 'force' option can bypass this failure, it replaces the
existing guideline with the imported one addressing a completely
different use-case.

There is no option to rename the imported guideline, which prevents a
successful import of backups when duplicate names occur.

This patch addresses that by adding the missing 'rename' option to the
.import_routing_guideline() command that allows renaming the imported
guideline when a name clash is detected.

Change-Id: I3041c8a07ac296e7ed8ffeb493e114bb3ffb0636