BUG#37866205 Relax the server's version check done by the dumper
As a fix for BUG#36701854, dumper compared the server's version with its
own, and threw an error if server had a greater minor version.

This check has been relaxed, an error is thrown only if server has a
greater major version, when the minor version is greater, a warning is
printed instead.

Change-Id: Ieb613ec2c04012b6715d293fbc1a40a34850ffb3