BUG#37867455 util.load_dump aborts when Performance Schema is not enabled
When loading a dump into an instance which had Performance Schema
disabled, operation would fail with:

  MySQL Error 1683 (HY000): 'ps_current_thread_id': The Performance Schema is not enabled.

The call to ps_current_thread_id() was introduced by a fix for
BUG#36197620.

As a fix, if call to ps_current_thread_id() fails for any reason, error
is logged and load continues.

Change-Id: Ia2e508d23eed141cf968ff41427bb06b08d2b77c