BUG#37669785 util.copyInstance should create users first
When loading a dump using an account which did not have the
ALLOW_NONEXISTENT_DEFINER privilage, operation would fail if dump
contained a non-view DDL object with a DEFINER clause set to an account
included in the dump. This happened, because non-view objects were
created before the accounts.

The account loading code has been reworked, it's now executed in several
steps, accounts are created before any other DDL objects (to make sure
that all accounts exist when statements with the DEFINER clause are
executed), while grants are applied just before view placeholders are
replaced with views (just as it was before).

Change-Id: Ie27f45996141c5de97af45891f4e21aa953dc30a