BUG#37979837 Custom SQL statements with comments are sent to server
Comments are meant to be ignored by the prefix matching logic used
to identify custom SQL from regular SQL, this is not the case when
using full line comments through the session.run_sql() API, as a
result, those statements were sent to the server for processing
always resulting in syntax error. This patch enhances the logic
in the prefix matching so full line comments are ignored.

Change-Id: I17fb3ce12224da98aead2d4d204133b4c2d0a491