BUG#38044015 Enable visibility of web only enabled plugin functions
Registering a python plugin function allows defining wether the function
should be available as a shell API, from the shell CLI interface or from
the web, however, it is not possible to define plugin functions that are
available on the web, but not in the shell, which is required to diable
the shell API functions of the MRS plugin, but keeping availability
from the GUI plugin to expose them to the VSCode Extension.

This patch enables that by keeping a registry of all the plugin function
definitions independently of their avilability configuration. Clients
could use this registry to made functions available independelty if they
are directly available on the shell or not.

Change-Id: I089b2b8a3e0dda9def080a872ee2d9a4288ea8be