BUG#36916939 Shell reads profiles without profile prefix
Shell read all profiles from the AWS config file, even the ones which
were not prefixed with the 'profile ' prefix.

In order to align the behaviour with the AWS CLI, non-default sections
of the AWS config file which are not prefixed with the 'profile '
prefix are ignored.

Change-Id: I5181d669f86e072d2748e52ac54d1c9427d9c767
(cherry picked from commit 37a3bfa3860be613231f409e14571675a142b8ed)