BUG#38107377 Copying instance fails with: version '9.3.1' is newer than the maximum version '9.3.0'
When using Shell 9.3.0 to copy instance to a MySQL Server 9.3.1,
operation failed with an error:

  Requested MySQL version '9.3.1' is newer than the maximum version
  '9.3.0' supported by this version of MySQL Shell (ArgumentError)

This error is reported by the validation code of the `targetVersion`
dump option, which is automatically set by the copy utilities to the
version of the target instance.

The validation has been relaxed, it now only compares the major.minor
version.

Additionally, the `ignoreVersion` copy option can now be used to
downgrade this error to a warning.

Change-Id: I85a27e448f5bb854ed4d14e991a1112f97e11c40