Fix AdminAPI tests affected by default value changes
WL#16438 changed the default value of
group_replication_recovery_use_ssl to 'ON' and
group_replication_ssl_mode to 'REQUIRED'.

We already check whether the target MySQL server supports SSL
before setting these variables to avoid failures, and that logic
remains unchanged. This check is still necessary because if the
server does not support SSL, and the default "ON" is left as-is,
Group Replication fails.

However, some tests required minor adjustments to account for
the new defaults introduced in 9.5.0.

Change-Id: I0977ff2026bd3b8c367662ae72d9f6f2258ca503