BUG#38224974 Enable strict mode when converting InnoDB-based vector store tables to Lakehouse
In order for the Lakehouse engine to support BASE64-encoded vector
columns, the guided load needs to be disabled. Previously, this was done
by disabling the strict mode, when creating the tables.

With this patch, the guided mode is disabled when executing the
SECONDARY_LOAD statement, by using the syntax introduced by WL#17062.
This allows to enable the strict mode.

Change-Id: Ic2c4ed9573feff21a3bd31076aefd41e9fa70486
(cherry picked from commit e1e237d395cc75280171d5336cdec308c5f34eb6)