Always write RPATH into bundled executables
With the fix for Bug#38261044, server codebase always sets INSTALL_RPATH
CMake property for all of its executables. This means that these
binaries are linked with the `rpath` linker option set to a dummy value,
which is replaced with the real one during the installation.

Given that Shell copies bundled binaries into its package, dummy RPATH
is not overwritten during the installation, and remains unchanged unless
we set the RPATH of our own. This only happened when OpenSSL was bundled
with the Shell's package, and caused RPM build errors when RPATH was not
altered.

In order to fix this issue, RPATH is always written into executables
bundled with Shell.

Change-Id: I129ee3aff2cd1ae1de974e39987793555a47abb4