BUG#38089433 Migration from MariaDB 11.8.2 fails: Unknown collation: 'utf8mb4_uca1400_ai_ci'
Loading a dump from a MariaDB 11.8.2 instance into a MySQL instance
failed with:
  Unknown collation: 'utf8mb4_uca1400_ai_ci'

MariaDB has a number of collations which are not available in MySQL, and
'utf8mb4_uca1400_ai_ci' is one of them. Recently it has become a default
collation for MariaDB installations.

As a fix, dumper detects unsupported collations, and replaces them with
closest compatible collation available in MySQL. A warning is emitted
for each replaced collation.

Change-Id: I00a944514224bf1679d22977813f326a815b6d51