BUG#38194922 non-standard FK check gets confused with duplicate indexes
This commit fixes a false posistive bug with foreign key check in Upgrade Checker in MySQL Shell.

When creating a foreign key constraint, it is required (from Server version 8.4.0) that referenced column has an unique index.
There is a possibility, that for the same table and column a second, non-unique key can be created.
Due to how the check queries the database, this produced invalid results, where check repoted the non unique index,
even when a unique one was present.

This commit fixes the check query, by limiting results, to show records for specific table and column only if
there is at least one non unique index and no other unique one.

Change-Id: I1eb5ba6fd8a8a8d0184c579f242771635227a4cb