BUG#38408207 Resuming an interrupted load causes duplicate entries in the progress file
When resuming a load which was interrupted while one of the chunks was
subchunked, loader skipped the subchunks which already had been loaded,
however it started counting new subchunks from zero. This meant that the
progress file contained multiple entries for i.e. subchunk 0, and if the
load was interrupted again, another resume would not skip the correct
number of bytes.

This has been fixed, when resuming an interrupted load, new subchunks
are assigned correct numbers, continuing numbering from the previous
load.

Change-Id: I71f8adc7a33e2b94a229f91064e9b5e5396e9095