BUG#37326937 load_dump should perform checks on dryRun
If loading a dump failed due to duplicate database objects, a second
attempt succeeded even though it should have failed as well.

The duplicate objects check is skipped when resuming a load, which is
determined by the presence of a non-empty progress file. The first
failed attempt created a progress file with some contents before
erroring out due to duplicate objects, and the second attempt did not
execute the check because the progress file existed.

As a fix, the progress file is now created just before loading actually
starts.

Change-Id: I444062fee89d94762d5f6db31086f625436af9f6