BUG#38442101 Shell crashes when running invalidEngineForeignKey check
This commit fixes a crash in MySQL Shell, when using Upgrade Checker with invalidEngineForeignKey check with tables containing non-letter characters (ex. "-").

Check fails to find a table in cache because of the escaping method that MySQL Server uses in information_schema.innodb_sys_foreign to encode table names.
This commit adds a step to decode common non-letter characters.

Change-Id: Iffb35ca46f14b2b4c912d7934870ab63406655cd