WL#17172 Upgrade Checker - filtering options
This patch introduces filtering options to the upgrade checker.
The following options are included:

- includeSchemas/excludeSchemas
- includeTables/excludeTables
- includeEvents/excludeEvents
- includeRoutines/excludeRoutines
- includeUsers/excludeUsers
- includeTriggers/excludeTriggers

If any include/exclude pair contains the same entry, an error will be
reported.

If both include/exclude pairs are defined, include takes priority.

It is not possible to include objects contained in a schema that was not
included or that was excluded.

It is not possible to include triggers on a table that was not included
or that was excluded.

Options added:
- includeSchema - filters in only specified schema identfiers when processing checks.
- excludeSchema - filters out specified schema identifiers when processing checks.

Change-Id: Ib8c3eaeacb92ab124740ddea5826d34bbf55a74f