Adding required typing annotations on plugin_function decorator.
The web parameter on the plugin_function decorator requires to support
either a boolean value to explicitly enable/disable the function on the
web layer. Additionally, a dictionary of options specific to the web
layer is supported, a non empty dictionary would implicitly enable the
function.

Additionally, a static validation is being done on the plugins for which
it is required that the web parameter reflects the allowed values.

Change-Id: Ibb2961d83b38b03db89fdd7c8132860cdf879032