BUG#38475999 Dump Silent Fails When Composite Unique Key Includes ENUM and Non-ENUM Columns
When dumping a table which has a composite key with an ENUM column which
is not the first one in the key definition, and whose values are not
listed alphabetically, Shell would not dump the complete table data.

Indexes with ENUM columns are no longer used to chunk the table data
(see BUG#35180061), however a misplaced break statement in a fix for
another BUG#36493316 has limitted the search for ENUM columns to the
first column in an index, ignoring the remaining ones. This has been
fixed.

Change-Id: Ifb20441910bc7f57ab96d1d799a243da9c465cd1