BUG#38566495 Load fails when resuming with 'dropExistingObjects' option set to true
When resuming a load with the 'dropExistingObjects' option set to true,
if the user executing the load lacked the `ALLOW_NONEXISTENT_DEFINER`
privilege, operation could have failed with:

  While dropping the account 'root'@'localhost': MySQL Error 4006
  (HY000): Operation DROP USER failed for 'root'@'localhost' as it is
  referenced as a definer account in a view.: DROP USER IF EXISTS
  'root'@'localhost'

Previously, when resuming the load, user accounts (unlike other DDL
objects whose status is tracked in the progress file) were always
dropped, regardless of whether they had been created in the previous
run.

With this fix, the creation of user accounts is logged in the progress
file. If the load is resumed, accounts that were previously created are
not dropped (if 'dropExistingObjects' option is enabled) and are not
created a second time.

Change-Id: Iebbf85f911f3178869b92db2e1c0414564a0102c